Refactor currentSoilStats into components
