+++ components/index.js
@@ -1,3 +1,4 @@
 export * from './Header';
-export * from './WeatherBox/WeatherBox';
+export * from './WeatherBox';
 export * from './CurrentWeather';
+export * from './DataCircle';
