+++ screens/Home/styles.js
@@ -48,17 +48,5 @@ export default StyleSheet.create({
   },
   currentWidgets: {
     flexDirection: 'row'
-  },
-  currentSoilStatContainer: {
-    alignItems: 'center',
-    alignItems: 'center',
-    backgroundColor: '#d5fdd5',
-    borderColor: '#A14550',
-    borderRadius: 30,
-    borderWidth: 2,
-    justifyContent: 'space-between',
-    margin: 3,
-    padding: 10,
-    width: 170
   }
 })
